Bug#36610878 uint16 overflow in `SEL_ARG::elements` leads to out-of-bounds writes
For really long in-lists, we may get out-of-bounds writes.
Change SEL_ROOT::elements from uint16 to size_t.

+++ sql/range_optimizer/tree.h
@@ -24,9 +24,10 @@
 #ifndef SQL_RANGE_OPTIMIZER_TREE_H_
 #define SQL_RANGE_OPTIMIZER_TREE_H_
 
-#include <assert.h>
-#include <string.h>
 #include <sys/types.h>
+#include <cassert>
+#include <cstddef>
+#include <cstring>
 
 #include "my_alloc.h"
 #include "my_base.h"
@@ -215,7 +216,7 @@ class SEL_ROOT {
   /**
     Number of nodes in the RB-tree, not including sentinels.
   */
-  uint16 elements{0};
+  size_t elements{0};
 };
 
 int sel_cmp(Field *f, uchar *a, uchar *b, uint8 a_flag, uint8 b_flag);
